Rename a file to match the name of the symbol exported from it
